+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.2.0 (not released)
   * Rapns.apns_feedback for one time feedback retrieval. Rapns.push no longer checks for feedback.
+  * Lazily connect to the APNs only when a notification is to be delivered.
 
 ## 3.1.0 (Jan 26, 2013)
   * Rapns.reflect API for fine-grained introspection.